prevent minimized applications from closing

+++ main.cpp
@@ -6,7 +6,7 @@ const int defaultDelay = 1;
 
 unsigned int *getActiveWindowsPID()
 {
-    CFArrayRef windows = CGWindowListCopyWindowInfo(kCGWindowListOptionOnScreenOnly, kCGNullWindowID);
+    CFArrayRef windows = CGWindowListCopyWindowInfo(kCGWindowListOptionOnScreenBelowWindow, kCGNullWindowID);
     unsigned int *pids = new unsigned int[CFArrayGetCount(windows)];
     pids[0] = CFArrayGetCount(windows);
 
@@ -23,6 +23,23 @@ unsigned int *getActiveWindowsPID()
 
     CFRelease(windows);
     return pids;
+}
+
+bool isWindowMinimized(unsigned int pid)
+{
+    AXUIElementRef appRef = AXUIElementCreateApplication(pid);
+    CFTypeRef minimizedValue;
+    bool isMinimized = false;
+
+    if (AXUIElementCopyAttributeValue(appRef, kAXWindowsAttribute, &minimizedValue) == kAXErrorSuccess)
+    {
+        CFIndex windowCount = CFArrayGetCount(static_cast<CFArrayRef>(minimizedValue));
+        isMinimized = windowCount > 0;
+        CFRelease(minimizedValue);
+    }
+
+    CFRelease(appRef);
+    return isMinimized;
 }
 
 int main()
@@ -45,7 +62,10 @@ int main()
             }
             if (!found)
             {
-                kill(previous[i], SIGTERM);
+                if (!isWindowMinimized(previous[i]))
+                {
+                    kill(previous[i], SIGTERM);
+                }
             }
         }
         delete[] previous;
